requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
opencv-python>=4.8.0
albumentations>=1.3.0
segmentation-models-pytorch>=0.3.0
tqdm>=4.65.0
scikit-learn>=1.3.0
gradio>=4.0.0
matplotlib>=3.7.0
pyyaml>=6.0
